Progress on refactoring tests
