+++ src/eval/mod.rs
@@ -2,5 +2,4 @@ pub mod emit_js;
 pub mod error;
 pub mod execution;
 pub mod helpers;
-pub mod tests;
 pub mod types;
