+++ src/lib.rs
@@ -1,2 +1,3 @@
 pub mod cmd;
 pub mod eval;
+pub mod tests;
